src/requirements.txt
psycopg2-binary==2.9.9
Scrapy==2.11.0
uvicorn==0.24.0.post1
fastapi==0.104.1
typing_extensions==4.9.0